requirements.txt
tensorflow==2.3.0
scikit-learn==0.22.1
matplotlib
python-dotenv
azureml-sdk
azure-cli
azureml-opendatasets==1.35.0    
azureml-core==1.35.0
azureml-telemetry~=1.35.0
